New setup working. Still need UI work.
NavTex messages are read from:
 - Cache file (json - schema hasn't been update to reflect SignalK's path format)
 - Serial as defined in plugin config (tty and baud rate)

The messages are send to SignalK in this format:
resources.navtex.B.A.42
{
  "epoch": 1644293523906,
  "text": [
    "121141 UTC JAN 22",
    "NORWEGIAN NAV. WARNING 29/2022",
    "CHART 61",
    "AREA TRAENA",
    "OVNHOLMSKJAERET LIGHT 66-32.2N 012-07.4E IS UNLIT."
  ]
}

The path resources.navtex.stationId.messageType.messageNr was choosen to allow easy filtering for the display based on stationId first and then message types.
messages numbers (00-99) are rotated through, so there should only be 1 message per number per stationId.

Epoch was used to make sorting based on time easy.

Text is an array as I didn't find a way to ensure the formatting would be kept as intended by the sender.